third_party/blink/web_tests/shadow-dom/v0/closed-mode-deep-combinator-and-shadow-pseudo-expected.txt

CONSOLE WARNING: line 306: /deep/ combinator is no longer supported in CSS dynamic profile. It is now effectively no-op, acting as if it were a descendant combinator. /deep/ combinator will be removed, and will be invalid at M65. You should remove it. See https://www.chromestatus.com/features/4964279606312960 for more details.
(1/6) /deep/ style rule on top-level document.
PASS backgroundColorOf('host_open_open') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1/div2') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3/div4') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5/div6') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7/div8') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7b') is "rgba(0, 0, 0, 0)"
(2/6) ::shadow style rule on top-level document.
PASS backgroundColorOf('host_open_open') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1/div2') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3/div4') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5/div6') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7/div8') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7b') is "rgba(0, 0, 0, 0)"
(3/6) /deep/ style on shadow tree.
PASS backgroundColorOf('host_open_open/div1') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1/div2') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1b') is "rgb(0, 0, 255)"
PASS backgroundColorOf('host_open_closed/div3') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3/div4') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3b') is "rgb(0, 0, 255)"
PASS backgroundColorOf('host_closed_open/div5') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5/div6') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5b') is "rgb(0, 0, 255)"
PASS backgroundColorOf('host_closed_closed/div7') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7/div8') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7b') is "rgb(0, 0, 255)"
(4/6) ::shadow style on shadow tree.
PASS backgroundColorOf('host_open_open/div1') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1/div2') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_open/div1b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3/div4') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_open_closed/div3b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5/div6') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_open/div5b') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7/div8') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host_closed_closed/div7b') is "rgba(0, 0, 0, 0)"
(5/6) /deep/ selector in querySelectorAll()
PASS host_open_open.querySelectorAll("div /deep/ div").length is 0
PASS host_open_closed.querySelectorAll("div /deep/ div").length is 0
PASS host_closed_open.querySelectorAll("div /deep/ div").length is 0
PASS host_closed_closed.querySelectorAll("div /deep/ div").length is 0
PASS results.length is 1
PASS node.id is "div1b"
PASS results.length is 1
PASS node.id is "div3b"
PASS results.length is 1
PASS node.id is "div5b"
PASS results.length is 1
PASS node.id is "div7b"
(6/6) ::shadow selector in querySelectorAll()
PASS host_open_open.querySelectorAll("div::shadow div").length is 0
PASS host_open_closed.querySelectorAll("div::shadow div").length is 0
PASS host_closed_open.querySelectorAll("div::shadow div").length is 0
PASS host_closed_closed.querySelectorAll("div::shadow div").length is 0
PASS div1.querySelectorAll("div::shadow div").length is 0
PASS div3.querySelectorAll("div::shadow div").length is 0
PASS div5.querySelectorAll("div::shadow div").length is 0
PASS div7.querySelectorAll("div::shadow div").length is 0
PASS successfullyParsed is true

TEST COMPLETE
